requirements_qwen.txt
# Python 3.10+ recommended
torch==2.7.1+cu118
torchvision==0.22.1+cu118
torchaudio==2.7.1+cu118
accelerate==1.10.1
transformers==4.57.1
qwen-vl-utils==0.0.11

# Image & video processing
Pillow==11.3.0
opencv-python-headless==4.12.0.88
decord==0.6.0
rembg==2.0.67
onnxruntime==1.23.1

# Optimization / model utils
einops==0.8.1
timm==1.0.22
safetensors==0.6.2
tokenizers==0.22.1
sentencepiece==0.2.1
flash_attn==2.8.3

# General utilities
numpy==2.0.1
tqdm==4.67.1
requests==2.32.5
huggingface-hub==0.34.4
psutil==7.0.0
protobuf==6.32.1
PyYAML==6.0.2

